Moved project around; added README and javadocs everywhere; minor refactors.
