--- experimental/retry/src/main/java/com/fnproject.fn.experimental/retry/DelayStrategy.java
@@ -1,7 +1,0 @@
-package com.fnproject.fn.experimental.retry;
-
-import java.io.Serializable;
-
-public interface DelayStrategy {
-    public long delayAmount(int attempt);
-}
