--- experimental/retry/src/main/java/com/fnproject.fn.experimental/retry/FullJitterStrategy.java
@@ -1,6 +1,0 @@
-package com.fnproject.fn.experimental.retry;
-
-import java.io.Serializable;
-
-public class FullJitterStrategy implements JitterStrategy, Serializable {
-}
