--- experimental/retry/src/main/java/com/fnproject.fn.experimental/retry/JitterStrategy.java
@@ -1,5 +1,0 @@
-package com.fnproject.fn.experimental.retry;
-
-public interface JitterStrategy {
-
-}
