--- experimental/retry/src/main/java/com/fnproject.fn.experimental/retry/StopStrategy.java
@@ -1,7 +1,0 @@
-package com.fnproject.fn.experimental.retry;
-
-import java.util.concurrent.TimeUnit;
-
-public interface StopStrategy {
-    public boolean shouldRetry(Throwable e, int attempt, int time, TimeUnit tu);
-}
